% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{add_leads_to_target_var}
\alias{add_leads_to_target_var}
\title{This helper function adds leads of target variable}
\usage{
add_leads_to_target_var(df, target_var_name, leads_vector)
}
\arguments{
\item{df}{raw data frame}

\item{target_var_name}{regression depended variable}

\item{leads_vector}{numeric vector with lead horizons}
}
\description{
This helper function adds leads of target variable
}
