% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimension_reduction_functions.R
\name{align_pca}
\alias{align_pca}
\title{Aligns the sign of PCA components}
\usage{
align_pca(pca_obj, var_name, positive_direction = TRUE)
}
\arguments{
\item{pca_obj}{PCA object}

\item{var_name}{string or numeric index. The name of the original
variable or it's column index in the data matrix}

\item{positive_direction}{a boolean indicator of the direction of the
original variable influence}
}
\value{
PCA object with aligned PCA matrix
 (x component of the PCA object list). The loadings are left unchaged.
}
\description{
This function aligns the PCA components with respect to chosen variable.
}
\details{
PCA assigns arbitrary sign to coefficients (loadings).
This function aligns the PCA components according to given increase
direction of the chosen original variable. The default direction is positive
}
