% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{calculate_CAGR}
\alias{calculate_CAGR}
\title{This function calculates the compound annualized growth rate}
\usage{
calculate_CAGR(df, horizon, freq = 4, forward = TRUE)
}
\arguments{
\item{df}{dataframe  - time indexed variable}

\item{horizon}{the horizon of the change period}

\item{freq}{time frequency of the data}

\item{forward}{determines whether the calculation is forward
(meaning that at each time point the change is between lead
 and current point) or backward (at each time point the change
 is between current point and its lag) looking.}
}
\description{
This function calculates the compound annualized growth rate
}
