% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{calculate_diff}
\alias{calculate_diff}
\title{This helper function calculates difference}
\usage{
calculate_diff(variable_vec)
}
\arguments{
\item{variable_vec}{data series}
}
\description{
This helper function calculates difference
}
