% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{calculate_four_quarters_ma}
\alias{calculate_four_quarters_ma}
\title{This helper function calculates 4 points moving average}
\usage{
calculate_four_quarters_ma(variable_vec)
}
\arguments{
\item{variable_vec}{data series}
}
\description{
This helper function calculates 4 points moving average
}
