% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing_functions.R
\name{calculate_implied_std_dev}
\alias{calculate_implied_std_dev}
\title{Calculate implied standard deviation}
\usage{
calculate_implied_std_dev(data_points_vec, percentiles_vec)
}
\arguments{
\item{data_points_vec}{vector of observed data points from
(assumed) normal distribution}

\item{percentiles_vec}{vector of percentiles corresponding to the
data points specified in \code{data_points_vec}}
}
\value{
the value of standard deviation of the normal (mean 0)
 distribution implied by the \code{data_points_vec} and
 \code{percentiles_vec} parameters
}
\description{
This function calculates the implied standard
deviation from normal distribution from one data point
}
