% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{calculate_percent_change}
\alias{calculate_percent_change}
\title{This helper function calculates percent_changes}
\usage{
calculate_percent_change(variable_vec)
}
\arguments{
\item{variable_vec}{data series}
}
\description{
This helper function calculates percent_changes
}
