% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxilary_functions.R
\name{calculate_skew_and_iqr}
\alias{calculate_skew_and_iqr}
\title{Calculate skew and IQR measures}
\usage{
calculate_skew_and_iqr(gar_obj, quantile_values = c("0.25", "0.5", "0.75"))
}
\arguments{
\item{gar_obj}{model object with run_GaR_analysis result}

\item{quantile_values}{vector that specifies the quantiles used in the calculation
of skew and iqr. The structure of the vector should be (low,mid,high), for example
if the iqr is based on 95th and 5th quantile the vector will be c(0.05,0.5,0.95).}
}
\description{
Calculate skew and IQR measures
}
\details{
This function takes gar object, extracts
predictions and calculates Skew and IQR measures
}
