% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calculate_win_len_from_date}
\alias{calculate_win_len_from_date}
\title{Calculate window length for start date

 @param start_date

 @param data_df}
\usage{
calculate_win_len_from_date(start_date, data_df)
}
\description{
This function determines the required \code{win_len}
 for the analysis set so that the evaluation set will starts at
 the specified date
}
