% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{calculate_yoy_changes}
\alias{calculate_yoy_changes}
\title{Calculate year on year percent changes}
\usage{
calculate_yoy_changes(variable_vec, data_frequency = "quarterly")
}
\arguments{
\item{variable_vec}{data series}

\item{data_frequency}{string that specifies time frequency of the data.

Available options are :
\itemize{
 \item {quarterly}{ (the default)}
 \item {monthly}
}}
}
\description{
This helper function calculates the percent change
between parallel periods in  two different years.
}
