% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{chain_index}
\alias{chain_index}
\title{This function creates a chained index partition}
\usage{
chain_index(df, preprocess_method = "pca", ...)
}
\arguments{
\item{df}{data frame with pca data}

\item{preprocess_method}{method for dimension reduction (default is PCA)}

\item{...}{external optional arguments}
}
\description{
This function creates a chained index partition
}
