% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{detrend.HP}
\alias{detrend.HP}
\title{This function filters non NA observations and detrends the clean
 series using HP filter}
\usage{
detrend.HP(data, HP_gamma = NULL)
}
\description{
@import mFilter
}
