% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxilary_functions.R
\name{extract_coeffs_from_gar_model}
\alias{extract_coeffs_from_gar_model}
\title{Extract regression coefficient from gar model object}
\usage{
extract_coeffs_from_gar_model(gar_model, partition_names = NULL)
}
\arguments{
\item{gar_model}{model object with run_GaR_analysis result}

\item{partition_names}{optional which partitions to return}
}
\value{
coeffs_df
}
\description{
This function extracts the coefficients data frame from gar model
}
