% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxilary_functions.R
\name{extract_factor_contribution_from_gar_model}
\alias{extract_factor_contribution_from_gar_model}
\title{Extract factor contribution from gar model}
\usage{
extract_factor_contribution_from_gar_model(gar_model, target_quantile = "0.05")
}
\arguments{
\item{gar_model}{model object with run_GaR_analysis result}

\item{target_quantile}{dplyr::filtering quantile (default 0.05)}
}
\value{
factor_contribution_df
}
\description{
Extract factor contribution from gar model
}
\details{
This function extracts the factor contribution (coefficients
 multiplied by values) data frame from gar model for given quantile
}
