% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxilary_functions.R
\name{extract_fitted_values_from_gar_model}
\alias{extract_fitted_values_from_gar_model}
\title{Extract fitted values data frame}
\usage{
extract_fitted_values_from_gar_model(gar_model)
}
\arguments{
\item{gar_model}{model object with run_GaR_analysis result}
}
\value{
data frame with columns date,horizon,quantile,fitted_values
}
\description{
This function extracts Pfitted values data frame
 from gar model
}
