% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxilary_functions.R
\name{extract_pca_loadings_from_gar_model}
\alias{extract_pca_loadings_from_gar_model}
\title{Extract PCA loadings}
\usage{
extract_pca_loadings_from_gar_model(gar_model)
}
\arguments{
\item{gar_model}{model object with run_GaR_analysis result}
}
\value{
pca_loadings_df
}
\description{
This function extracts PCA loadings data frame from gar model
}
