% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxilary_functions.R
\name{extract_pca_timeseries_from_gar_model}
\alias{extract_pca_timeseries_from_gar_model}
\title{Extract PCA timeseries}
\usage{
extract_pca_timeseries_from_gar_model(gar_model, n_comp = 1)
}
\arguments{
\item{gar_model}{model object with run_GaR_analysis result}

\item{n_comp}{number of PCA components to return}
}
\value{
pca_timeseries_df
}
\description{
This function extracts PCA timeseries data frame from gar model
}
