% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{extract_preprocess_arguments}
\alias{extract_preprocess_arguments}
\title{Extract preprocess arguments}
\usage{
extract_preprocess_arguments(partitions_list, target_var_name = NULL)
}
\arguments{
\item{partitions_list}{a list of partitions for dimension reduction.
For elements in partition that contain only one variable the variable returns "as is".}
}
\description{
This is an auxiliary function for preprocess_df
}
