% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxilary_functions.R
\name{extract_qreq_coeff_table}
\alias{extract_qreq_coeff_table}
\title{Extract coefficients from quantile reg}
\usage{
extract_qreq_coeff_table(qreg_obj)
}
\arguments{
\item{qreg_obj}{quantile regression object}
}
\description{
This function extracts the coefficients from quantile regression
}
