% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing_functions.R
\name{extract_smoothed_quantiles}
\alias{extract_smoothed_quantiles}
\title{This function returns "smoothed" (drawn from parametrized t skew distribution)
quantiles}
\usage{
extract_smoothed_quantiles(
  dist_param_df,
  raw_quantiles_df,
  smoothed_quantiles_vec_x = c(0.05, 0.25, 0.5, 0.75, 0.95)
)
}
\arguments{
\item{dist_param_df}{data frame that specifies distribution parameters.
The structure is (date, horizon, t_skew_params, values)}

\item{raw_quantiles_df}{data frame that specifies t skew fitted quantiles
The structure is (date, horizon, quantile, values)}

\item{smoothed_quantiles_vec_x}{vector of required quantiles.
Default is c(0.05, 0.25, 0.5, 0.75, 0.95)}
}
\description{
This function returns "smoothed" (drawn from parametrized t skew distribution)
quantiles
}
