% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{fill_na_average}
\alias{fill_na_average}
\title{Fill na with average of k previous observations}
\usage{
fill_na_average(data_vec, k = 4)
}
\arguments{
\item{data_vec}{vector of data}

\item{k}{window size}
}
\description{
Fill na with average of k previous observations
}
