% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing_functions.R
\name{fit_t_skew}
\alias{fit_t_skew}
\title{Fit t skewed distribution}
\usage{
fit_t_skew(
  estimated_df,
  time_limit = 10,
  bounded_optimization = TRUE,
  lower_bounds = c(-Inf, 0.01, -Inf, 1),
  upper_bounds = c(Inf, Inf, Inf, 100)
)
}
\arguments{
\item{estimated_df}{data frame with (quantiles, values) columns}

\item{time_limit}{the time limit given to each optimization round.
In case the computation doesn't converge in the given time frame a vector
of 0 is returned for the optimization parameters and "timed out" warning is
issued.}

\item{bounded_optimization}{boolean indicator, default TRUE}

\item{lower_bounds}{lower bounds for xi (location),
omega (scale), alpha (slant), nu (degrees of freedom).
The default is (-Inf, 0.01, -Inf, 0).}

\item{upper_bounds}{upper bounds for xi (location),
omega (scale), alpha (slant), nu (degrees of freedom).
The default is (Inf, Inf, Inf, 100).}
}
\value{
vector of parameters xi (location),
omega (scale), alpha (slant), nu (degrees of freedom) if the fitting was
 successful, otherwise a vector of 0 if the optimization has timeout or a
 vector of NA if there was an error.
}
\description{
This function fits t skewed distribution based on empirical
quantiles. The input data should be a data frame with (quantiles, values)
columns.
}
\details{
The fitting is performed by calculating the errors
(difference between empirical and theoretical values for each quantile).
The loss metric is RMSE.
}
