% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing_functions.R
\name{fit_t_skew_to_gar_df}
\alias{fit_t_skew_to_gar_df}
\title{Fit t skewed distribution to gar df}
\usage{
fit_t_skew_to_gar_df(
  gar_df,
  return_smoothed_quantiles = FALSE,
  smoothed_quantiles_vec = c(0.05, 0.25, 0.5, 0.75, 0.95),
  time_limit = 10,
  bounded_optimization = TRUE,
  lower_bounds = c(-Inf, 0.01, -Inf, 1),
  upper_bounds = c(Inf, Inf, Inf, 100),
  parallel_computing = FALSE
)
}
\arguments{
\item{gar_df}{data frame. The result of run_GaR_analysis}

\item{return_smoothed_quantiles}{boolean returns quantiles}

\item{smoothed_quantiles_vec}{vector of required quantiles.
Default is c(0.05, 0.25, 0.5, 0.75, 0.95)}

\item{time_limit}{the time limit given to each optimization round.
In case the computation doesn't converge in the given time frame a vector
of 0 is returned for the optimization parameters and "timed out" warning is
issued.}

\item{bounded_optimization}{boolean indicator, default TRUE}

\item{lower_bounds}{lower bounds for xi (location),
omega (scale), alpha (slant), nu (degrees of freedom).
The default is (-Inf, 0, -Inf, 0).}

\item{upper_bounds}{upper bounds for xi (location),
omega (scale), alpha (slant), nu (degrees of freedom).
The default is (Inf, Inf, Inf, 100).}

\item{parallel_computing}{boolean. If TRUE uses parallel computing
by setting \code{plan(multisession)} before fitting t skewed distribution
and setting \code{plan(sequential)} after the fitting is complete.}
}
\description{
This function fits t skewed distribution to either fitted
 values of gar model or out of sample forecasted values
}
