% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{fix_quantile_crossing}
\alias{fix_quantile_crossing}
\title{Fixes quantile regression crossing}
\usage{
fix_quantile_crossing(prediction_df)
}
\arguments{
\item{prediction_df}{data frame with prediction values. The structure is
date, horizon, quantile, forecast_values or fitted_values}
}
\description{
This function fixes the crossing issue in quantile
 regression by reordering the values according to quantiles order
}
