% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_validation_functions.R
\name{get_gar_forecast}
\alias{get_gar_forecast}
\title{Out of sample forecasts}
\usage{
get_gar_forecast(
  partitions_list,
  vars_df,
  target_var_name,
  horizon_list,
  quantile_vec,
  pca.align.list = NULL,
  preprocess_method = "pca",
  win_len = 30,
  win_type_expanding = TRUE
)
}
\arguments{
\item{partitions_list}{list of partition names}

\item{vars_df}{data frame with input variables}

\item{target_var_name}{string that specifies outcome variable}

\item{horizon_list}{list (or vector) of forecast horizons}

\item{quantile_vec}{vector quantiles for quantile regression}

\item{pca.align.list}{(optional) list list that contains the "aligning" variable
for each partition that should be "aligned". See}

\item{preprocess_method}{dimension reduction method (default PCA)}

\item{win_len}{the length of sliding window (default 30).
Can be supplied as start date of the analysis (test) set}

\item{win_type_expanding}{boolean should the sliding window expand
(default TRUE)}
}
\value{
a data frame with out-of-sample predictions.
The "date" column specifies the date of the last observation in the test set,
 namely, the date of the last observation used to construct the forecast.
 The "forecast_target_date" specifies the date for which the forecast is
  aimed, namely, forecast_target_date = date + horizon.
}
\description{
This is a convenient function that calculates out of
sample forecast.
}
\details{
An out of sample forecast is calculated by rolling
regression (determined by \code{win_len}) that forecast
ahead according to \code{out_of_sample_step}
}
