% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_partition_combs}
\alias{get_partition_combs}
\title{Return combinations of partition elements}
\usage{
get_partition_combs(partitions_list, partition_name)
}
\arguments{
\item{partitions_list}{a named list where the names are
\itemize{
 \item{"optional"}{-Optional elements}
 \item{"required"}{-Required elements}
 }}

\item{partition_name}{name of the partition for resulting df}
}
\description{
This function returns a data_frame with list-column
partition combs
}
