% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing_functions.R
\name{get_t_skew_quantiles}
\alias{get_t_skew_quantiles}
\title{This is an auxiliary function that returns quantiles from specified
t skew distribution}
\usage{
get_t_skew_quantiles(
  t_skew_param_df,
  quantiles_vec = c(0.05, 0.25, 0.5, 0.75, 0.95)
)
}
\arguments{
\item{t_skew_param_df}{data frame that specifies distribution parameters.
The structure is (date, horizon, t_skew_params, values)}

\item{quantiles_vec}{vector of required quantiles.
Default is c(0.05,0.25,0.5,0.75,0.95)}
}
\description{
This is an auxiliary function that returns quantiles from specified
t skew distribution
}
