% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{get_time_indices_list}
\alias{get_time_indices_list}
\title{Get list of unique time indices of non missing data}
\usage{
get_time_indices_list(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
This function gets the time index for non missing
data for each variable in the data set and returns a list
of unique time indices
}
