% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{identify.endpoints.NA}
\alias{identify.endpoints.NA}
\title{This function identifies consequitive NA sequences
at the beginning or end of data vector. This is an auxilary function
for interpolate}
\usage{
\method{identify}{endpoints.NA}(data_vec)
}
\arguments{
\item{data_vec}{}
}
\description{
This function identifies consequitive NA sequences
at the beginning or end of data vector. This is an auxilary function
for interpolate
}
