% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{identify_endpoints_NA}
\alias{identify_endpoints_NA}
\title{This function identifies consequitive NA sequences
at the beginning or end of data vector. This is an auxilary function
for interpolate}
\usage{
identify_endpoints_NA(data_vec)
}
\arguments{
\item{data_vec}{data series}
}
\description{
This function identifies consequitive NA sequences
at the beginning or end of data vector. This is an auxilary function
for interpolate
}
