% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{identify_frequency}
\alias{identify_frequency}
\title{Identify time frequency of date column}
\usage{
identify_frequency(date_col)
}
\arguments{
\item{date_col}{}
}
\description{
This function identifies the time frequency of a
 date column in a data frame
}
