% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.commodities}
\alias{import.commodities}
\title{This function imports energy prices data from commodities file}
\usage{
import.commodities(filepath = NULL)
}
\description{
This function imports energy prices data from commodities file
}
