% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.gdp_without_startup_tax}
\alias{import.gdp_without_startup_tax}
\title{This function imports real short rate from real rate 1Y file}
\usage{
import.gdp_without_startup_tax(filepath = NULL)
}
\description{
This function imports real short rate from real rate 1Y file
}
