% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.global.monetary.rates}
\alias{import.global.monetary.rates}
\title{This function imports global monetary rates data from csv format}
\usage{

  import.global.monetary.rates(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\Data\\\\BOI\\\\",
  "DSGE\\\\global monetary rates.csv"))
}
\description{
This function imports global monetary rates data from csv format
}
