% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.global_gdp}
\alias{import.global_gdp}
\title{This function imports global economic data from global_gdp file}
\usage{

  import.global_gdp(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\",
  "Data\\\\BOI\\\\DSGE\\\\", "global gdp.csv"))
}
\description{
This function imports global economic data from global_gdp file
}
