% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.global_inflation}
\alias{import.global_inflation}
\title{This function imports global economic data from global_gdp file}
\usage{
import.global_inflation(filepath = NULL)
}
\description{
This function imports global economic data from global_gdp file
}
