% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.iip.data}
\alias{import.iip.data}
\title{This function imports IIP data from xls format}
\usage{
import.iip.data(filepath = NULL)
}
\description{
This function imports IIP data from xls format
}
