% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.inflation_sa}
\alias{import.inflation_sa}
\title{This function imports CPI data from inflation_sa file}
\usage{

  import.inflation_sa(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\Data",
  "\\\\BOI\\\\DSGE\\\\inflation_sa.xlsx"))
}
\description{
This function imports CPI data from inflation_sa file
}
