% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.Inputs_financial}
\alias{import.Inputs_financial}
\title{This function imports current account data from Inputs_CurrentAccount file}
\usage{

  import.Inputs_financial(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\Data",
  "\\\\BOI\\\\DSGE\\\\Inputs_financial.csv"))
}
\description{
This function imports current account data from Inputs_CurrentAccount file
}
