% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.Inputs_fw_row}
\alias{import.Inputs_fw_row}
\title{This function imports yields data from Inputs_fw_row file}
\usage{
import.Inputs_fw_row(filepath = NULL)
}
\description{
This function imports yields data from Inputs_fw_row file
}
