% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.Inputs_labor}
\alias{import.Inputs_labor}
\title{This function imports  labor market data from Inputs_labor file}
\usage{

  import.Inputs_labor(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\Data",
  "\\\\BOI\\\\DSGE\\\\Inputs_labor.xlsx"))
}
\description{
This function imports  labor market data from Inputs_labor file
}
