% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.Inputs_NA_SA}
\alias{import.Inputs_NA_SA}
\title{This function imports National Accounts data from Inputs_NA_SA file}
\usage{
import.Inputs_NA_SA(filepath = NULL)
}
\description{
This function imports National Accounts data from Inputs_NA_SA file
}
