% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.michael.data}
\alias{import.michael.data}
\title{This function imports all Michael's data}
\usage{
import.michael.data(filepath = NULL)
}
\description{
This function imports all Michael's data
}
