% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.monetary_variables}
\alias{import.monetary_variables}
\title{This function imports monetary data from monetary_variables file}
\usage{

  import.monetary_variables(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\Data",
  "\\\\BOI\\\\DSGE\\\\monetary variables.xlsx"))
}
\description{
This function imports monetary data from monetary_variables file
}
