% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.real_rate_1Y}
\alias{import.real_rate_1Y}
\title{This function imports real short rate from real rate 1Y file}
\usage{
import.real_rate_1Y(filepath = NULL)
}
\description{
This function imports real short rate from real rate 1Y file
}
