% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.tourists}
\alias{import.tourists}
\title{This function imports tourists data from tourists file}
\usage{
import.tourists(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents",
  "\\\\Data", "\\\\BOI\\\\DSGE\\\\tourists.csv"))
}
\description{
This function imports tourists data from tourists file
}
