% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import_dsge_forecasts}
\alias{import_dsge_forecasts}
\title{Import DSGE forecasts}
\usage{
import_dsge_forecasts(file_path, quantiles = c(0.05, 0.25, 0.5, 0.75, 0.95))
}
\arguments{
\item{file_path}{string}

\item{quantiles}{numeric vector that specifies the required
quantiles for the DSGE forecast}
}
\description{
This function imports DSGE forecasts and convert the data
to tidy format.
}
\details{
This function imports staff forecast df
}
