% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import_from_fame_template}
\alias{import_from_fame_template}
\title{Import data from fame template}
\usage{
import_from_fame_template(template_path, data_frequency = "quarterly")
}
\arguments{
\item{template_path}{file path to template file}

\item{data_frequency}{string that specifies time frequency of the data.

Available options are :
\itemize{
 \item {quarterly}{ (the default)}
 \item {monthly}
}}
}
\value{
tidy df
}
\description{
This function imports and cleans data from fame template
}
\details{
If a complimentary column (a column name with _append suffix) is
 supplied, the functions merges the two columns (varname, varname_append)
 into one. The merge is performed by substituting NA values in the main
 (varname) column with values from the complimentary (varname_append) column.
}
