% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_function.R
\name{interpolate}
\alias{interpolate}
\title{This function interpolates forward by using linear interpolation}
\usage{
interpolate(data_vec, direction = "forward")
}
\arguments{
\item{data_vec}{}
}
\description{
This function interpolates forward by using linear interpolation
}
