% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{is_partition_identical}
\alias{is_partition_identical}
\title{compare two partitions}
\usage{
is_partition_identical(source_partition, target_partition)
}
\arguments{
\item{source_partition}{benchmark partition}

\item{target_partition}{compared partition}
}
\description{
This function compares two partitions
}
