% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_functions.R
\name{make_prediction_df}
\alias{make_prediction_df}
\title{Make prediction df}
\usage{
make_prediction_df(gar_model, xreg_df)
}
\arguments{
\item{gar_model}{model object with run_GaR_analysis result}

\item{xreg_df}{xreg data}
}
\description{
This function returns a data frame with predicted values
}
\details{
The default is in sample prediction (fitted values),
otherwise predict according to supplied xreg data
}
