% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{make_quant_reg_df}
\alias{make_quant_reg_df}
\title{This function creates a data set for quantile regression}
\usage{
make_quant_reg_df(
  vars_df,
  target_var_name,
  horizon_list,
  preprocess_method = "pca",
  partitions_list = NULL,
  pca.align.list = NULL,
  return_objects_list = FALSE
)
}
\arguments{
\item{vars_df}{data frame with input variables}

\item{target_var_name}{string that specifies outcome feature}

\item{horizon_list}{list of forecast horizon}

\item{preprocess_method}{string a method that aggregates the data to partitions}

\item{partitions_list}{a list of partitions for dimension reduction.
For elements in partition that contain only one variable the variable returns "as is".}

\item{pca.align.list}{named list where the names are partition names and the
first element is the aligning variable name. The second (optional) element
is the alignment direction (default is "positive")}

\item{return_objects_list}{boolean indicator that returns PCA objects.}
}
\value{
regression data frame
}
\description{
This function creates a data set for quantile regression
}
\details{
This functions applies preprocessing by reducing dimension of the data
(currently by PCA or PLS).
}
