% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimension_reduction_functions.R
\name{map_pca_reduction}
\alias{map_pca_reduction}
\title{Map PCA over partitions list}
\usage{
map_pca_reduction(
  multi_feature_partitions,
  vars_df,
  n_components = 1,
  pca_align_list = NULL
)
}
\arguments{
\item{multi_feature_partitions}{list of partitions}

\item{vars_df}{raw data frame}

\item{n_components}{number of components (default is 1)}

\item{pca_align_list}{(optional) a named list of alignment parameters.
The name is the name of the targeted partition. The first element in
the list is the aligning (axis) variable, the value is either character
(variable's name) or numeric (variable's position index). The second element
(optional) is boolean indicator alignment direction (True means
 positive direction).}
}
\description{
Applies dimension reduction using PCA to each partition in
partitions list
}
\details{
`align_pca` When pca coefficients vector doesn't contain the aligning
variable (perhaps because of name misspecification) a warning is issued
 and no alignment is performed.
}
