% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimension_reduction_functions.R
\name{map_pls_reduction}
\alias{map_pls_reduction}
\title{Map PLS reduction over partitions list}
\usage{
map_pls_reduction(
  multi_feature_partitions,
  vars_df,
  target_var_name,
  n_components = 1
)
}
\arguments{
\item{multi_feature_partitions}{list of partitions}

\item{vars_df}{raw data frame}

\item{target_var_name}{string that specifies outcome feature}

\item{n_components}{number of components (default is 1)}
}
\description{
Map PLS reduction over partitions list
}
