% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization_functions.R
\name{plot_fan_chart}
\alias{plot_fan_chart}
\title{This function produces the fan chart plots}
\usage{
plot_fan_chart(forecast_df, actual_df, fan_chart_date = NULL)
}
\arguments{
\item{forecast_df}{data frame with forecasted values.
by horizon, quantile, forecast_target_date, forecast_values.
The forecast_target_date specifies the date for which the forecast is aimed}

\item{actual_df}{data frame with actual values
by date and actual_values}

\item{fan_chart_date}{a date of the initial time point in the fan chart.
The default is the last date of forecast df}
}
\description{
This function produces the fan chart plots
}
