% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimension_reduction_functions.R
\name{pls_reduction}
\alias{pls_reduction}
\title{PLS reduction}
\usage{
pls_reduction(df, target_var_name, center = TRUE, scale = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{target_var_name}{string that specifies outcome feature}

\item{center}{boolean indicator}

\item{scale}{boolean indicator}
}
\value{
a list with two elements : pls_obj - PLS object,
 time_index - dates vector
}
\description{
PLS reduction
}
\details{
This function reduces dimension based on PLS method
The function takes a data matrix and returns the first n_comps
components of PLS transformation. If the data matrix has a
time index the result is aligned along the index
}
