% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{preprocess_df}
\alias{preprocess_df}
\title{Preprocess raw data}
\usage{
preprocess_df(
  df,
  vars_to_yoy = NULL,
  vars_to_percent_changes = NULL,
  vars_to_diff = NULL,
  vars_to_4_ma = NULL,
  convert_to_percent_units = FALSE
)
}
\arguments{
\item{df}{raw data frame}

\item{vars_to_yoy}{(optional) vector of variable names
for "Year on Year" transformation. Computes the percent change
between parallel periods in  two different years.}

\item{vars_to_percent_changes}{(optional) vector of variable names
Computes the percent change between two sequential  periods.}

\item{vars_to_diff}{(optional) vector of variable names for differencing transformation}

\item{vars_to_4_ma}{(optional) vector of variable names for "Annual moving average"
transformation}

\item{convert_to_percent_units}{logical default is FALSE should the relative change
variables be converted to percent units (multiply by 100).}
}
\description{
This function performs preprocessing variable transformation
}
\details{
The following transformations are supported
\itemize{
 \item{Year on Year percent change}
 \item{Differencing}
 \item{Annual moving average (the variables should be at quarterly frequency)}
}

In case of Year on Year percent change the function identifies the time
frequency (by checking date class, either yearqtr or yearmon) and calculates
the percent change appropriately
}
