% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goodness_of_fit_functions.R
\name{quantile_pit_score}
\alias{quantile_pit_score}
\title{Calculate PIT score}
\usage{
quantile_pit_score(forecast_df, actual_df)
}
\arguments{
\item{forecast_df}{data frame with forecasted values.
by horizon, quantile, forecast_target_date, forecast_values.
The forecast_target_date specifies the date for which the forecast is aimed}

\item{actual_df}{data frame with actual values
by date and actual_values}
}
\description{
Calculate PIT score
}
\details{
This function calculates the Probability Integral
Transformation to evaluate goodness of fit.
The value represents the calibration - the relative frequency of the
data points in the sample.

The evaluation is based on the assumption that the date in
\code{forecast_df} refers to the time in which the forecast was performed.
Namely, the function offsets each forecast date in \code{forecast_df} by the
relevant horizon and matches it with the respective date in \code{actual_df}.
For example, a forecast for the horizon of 4 quarters in 1999 Q1 is compared
to an actual value in 2000 Q1.
}
