% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goodness_of_fit_functions.R
\name{quantile_prediction_score}
\alias{quantile_prediction_score}
\title{Calculate prediction score}
\usage{
quantile_prediction_score(forecast_dist_df, actual_df)
}
\arguments{
\item{forecast_dist_df}{data frame with estimated t skewed distribution
by horizon, quantile, date and parameter}

\item{actual_df}{data frame with actual values
by value and date}
}
\description{
Calculate prediction score
}
\details{
This function calculates the prediction score which is the
probability to get the observed (actual) value from the fitted t skew
distribution based on estimated quantiles
}
