% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Goodness_of_fit_functions.R
\name{quantile_r2_score_calculation}
\alias{quantile_r2_score_calculation}
\title{Calculate quantile R square score for given spec:
realized_values, forecast_values, quantile, benchmark_values}
\usage{
quantile_r2_score_calculation(
  realized_values,
  forecast_values,
  quantile,
  benchmark_values
)
}
\arguments{
\item{realized_values}{realized values series for comparison to forecasted}

\item{forecast_values}{forecasted values series for comparison to realized}

\item{quantile}{the quantile corresponding to the realized and forecasted
values series}

\item{benchmark_values}{benchmark_values series (intercept predictions
in GaR model)}
}
\description{
This function evaluates the goodness of fit between
"realized"  value and forecasted quantiles. The calculation is for given spec:
realized_values, forecast_values, quantile, benchmark_values
}
