% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimension_reduction_functions.R
\name{reduce_data_dimension}
\alias{reduce_data_dimension}
\title{Data dimension reduction}
\usage{
reduce_data_dimension(
  vars_df,
  partition_list,
  target_var_name = NULL,
  n_components = 1,
  pca_align_list = NULL,
  preprocess_method = "pca",
  return_objects_list = FALSE
)
}
\arguments{
\item{vars_df}{a dataframe with all variables}

\item{partition_list}{a list of partitions for dimension reduction.
For elements in partition that contain only one variable the variable returns "as is".}

\item{target_var_name}{string that specifies outcome feature}

\item{n_components}{number of components that should be returned}

\item{pca_align_list}{(optional) a named list of alignment parameters.
The name is the name of the targeted partition. The first element in
the list is the aligning (axis) variable, the value is either character
(variable's name) or numeric (variable's position index). The second element
(optional) is boolean indicator alignment direction (True means
 positive direction).}

\item{preprocess_method}{(optional) string that specifies preprocess method
(default is pca)}

\item{return_objects_list}{boolean indicator that specifies whether a list
with object containing information (such as loadings) should be returned}
}
\value{
a list where first element is regression data (named xreg) and second
(optional) element is the pca obj list
}
\description{
This function preprocess data by reducing dimension and returns regression dataset
}
