% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_validation_functions.R
\name{run_cross_validation}
\alias{run_cross_validation}
\title{Run cross validation for quantile regression}
\usage{
run_cross_validation(
  partitions_list,
  vars_df,
  target_var_name,
  horizon_list,
  quantile_vec,
  pca.align.list = NULL,
  preprocess_method = "pca",
  win_len = 30,
  win_type_expanding = TRUE,
  ...
)
}
\arguments{
\item{partitions_list}{list of partition names}

\item{vars_df}{data frame with input variables}

\item{target_var_name}{string that specifies outcome variable}

\item{horizon_list}{list (or vector) of forecast horizons}

\item{quantile_vec}{vector quantiles for quantile regression}

\item{pca.align.list}{(optional) list list that contains the "aligning" variable
for each partition that should be "aligned". See}

\item{preprocess_method}{dimension reduction method (default PCA)}

\item{win_len}{the length of sliding window (default 30).
Can be supplied as start date of the analysis (test) set}

\item{win_type_expanding}{boolean should the sliding window expand
(default TRUE)}
}
\description{
The function performs rolling ("walk forward")
quantile regression.
}
\details{
After the estimation of the parameters on
the rolling window sample prediction for last observation is performed.
This is and "out of sample" prediction since the last observation data was
not used in the estimation of the parameters (due to unknown target feature)
}
