% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_functions.R
\name{run_GaR_analysis}
\alias{run_GaR_analysis}
\title{Perform GaR analysis}
\usage{
run_GaR_analysis(
  partitions_list,
  vars_df,
  target_var_name,
  horizon_list,
  quantile_vec,
  preprocess_method = "pca",
  pca.align.list = NULL,
  return_objects_list = TRUE
)
}
\arguments{
\item{partitions_list}{list of partitions}

\item{vars_df}{data frame with input variables}

\item{target_var_name}{string. that specifies outcome feature}

\item{horizon_list}{list of forecast horizon(s)}

\item{quantile_vec}{vector of required quantiles in quantile
regression (corresponds to tau argument in rq)}

\item{preprocess_method}{string. a method that aggregates the data
to partitions \cr
Possible options are:
\itemize{
 \item {pca}{ (the default)}
 \item {pls}
}}

\item{pca.align.list}{(Optional) A list that specifies the PCA aligning variable for
each partition and alignment direction (default is positive direction).
The first element in the list is the aligning (axis) variable, the
value is either character (variable's name) or numeric
(variable's position index). The second element if(supplied)
is a boolean indicator of alignment direction
(True means positive direction).
If no list is supplied then no alignment takes place}

\item{return_objects_list}{boolean indicator that returns PCA objects.
Default is TRUE}
}
\value{
reg_df data frame with training set data

qreg_result a quantile regression model object

fitted_df a data frame with fitted values

pca_obj (optional) PCA object
}
\description{
This is a convenient function that encapsulates the GaR analysis.
The stages of the analysis are :
\itemize{
  \item Perform dimension reduction by aggregating features
   to partitions (default method is PCA)
  \item Run quantile regression
}
}
