% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_functions.R
\name{run_quant_reg}
\alias{run_quant_reg}
\title{This function runs quantile regression}
\usage{
run_quant_reg(
  reg_df,
  target_var_name,
  quantile_vec,
  horizon_list,
  reg_type = "quantile",
  ...
)
}
\arguments{
\item{reg_df}{data for quantile regression}

\item{target_var_name}{string}

\item{quantile_vec}{vector of quantiles}

\item{horizon_list}{list of target horizons}

\item{reg_type}{type of regression (default "quantile").
The possible options are:
\itemize{
 \item {quantile}{ (the default)}
 \item {lasso}
}}

\item{...}{external optional arguments}
}
\value{
list of quantile reg objects
}
\description{
This function runs quantile regression
}
