% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing_functions.R
\name{run_t_skew_optimization}
\alias{run_t_skew_optimization}
\title{Run t skewed optimization}
\usage{
run_t_skew_optimization(
  estimated_df_x,
  bounded_optimization_x = TRUE,
  lower_bounds_x = c(-Inf, 0.01, -Inf, 1),
  upper_bounds_x = c(Inf, Inf, Inf, 100)
)
}
\arguments{
\item{estimated_df_x}{data frame with (quantiles, values) columns}

\item{bounded_optimization_x}{boolean indicator, default TRUE}

\item{lower_bounds_x}{lower bounds for xi (location),
omega (scale), alpha (slant), nu (degrees of freedom).
The default is (-Inf, 0, -Inf, 1).}

\item{upper_bounds_x}{upper bounds for xi (location),
omega (scale), alpha (slant), nu (degrees of freedom).
The default is (Inf, Inf, Inf, 100).}
}
\description{
This function finds the parameters of the t skewed distribution
xi (location), omega (scale), alpha (slant) and nu (degrees of freedom)
while minimizing the RMSE loss function. The function supports bounded and
unbounded optimization, the default is bounded optimization.
}
