% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoothing_functions.R
\name{t_skew_loss}
\alias{t_skew_loss}
\title{Calculate t skewed distribution fitting loss}
\usage{
t_skew_loss(estimated_df, t_skew_params)
}
\arguments{
\item{estimated_df}{data frame with (quantiles, values) columns}

\item{t_skew_params}{the parameters of the t skewed distribution
xi (location), omega (scale), alpha (slant) and nu (degrees of freedom)}
}
\description{
This function calculates the loss of matching estimated quantiles
with theoretical ones. The errors are calculated as the difference between
empirical and theoretical values for each quantile. The loss metric is RMSE.
}
