% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{warn_about_target_var_suffix}
\alias{warn_about_target_var_suffix}
\title{Validate compatability between target and processed variables}
\usage{
warn_about_target_var_suffix(target_var_name, partitions_list)
}
\description{
This function compares target variable name to partitions list
vars and issues a warning in case of unprocessed target var and
processed explanatory var
}
