% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{extract_preprocess_arguments}
\alias{extract_preprocess_arguments}
\title{Extract preprocess arguments}
\usage{
extract_preprocess_arguments(partitions_list, target_var_name = NULL)
}
\arguments{
\item{partitions_list}{a list of partitions for dimension reduction.
For elements in partition that contain only one variable the variable returns "as is".

 The function also returns unchanged variables names (storing them under
 the \code{asis_vars} name in transformations list that
 is passed to \code{preprocessed_df})}
}
\description{
This is an auxiliary function for preprocess_df
}
