% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.commodities}
\alias{import.commodities}
\title{This function imports energy prices data from commodities file}
\usage{

  import.commodities(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents",
  "\\\\Data\\\\BOI\\\\DSGE", "\\\\commodities.csv"))
}
\description{
This function imports energy prices data from commodities file
}
