% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.indus.prod.data}
\alias{import.indus.prod.data}
\title{This function imports Industrial Production data from csv format}
\usage{

  import.indus.prod.data(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\",
  "Data\\\\BOI\\\\DSGE\\\\", "Industrial production.csv"))
}
\description{
This function imports Industrial Production data from csv format
}
