% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.Inputs_CurrentAccount}
\alias{import.Inputs_CurrentAccount}
\title{This function imports current account data from Inputs_CurrentAccount file}
\usage{
import.Inputs_CurrentAccount(filepath = NULL)
}
\description{
This function imports current account data from Inputs_CurrentAccount file
}
