% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.inputs_foreign}
\alias{import.inputs_foreign}
\title{This function imports world economic data from inputs_foreign file}
\usage{

  import.inputs_foreign(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\Data",
  "\\\\BOI\\\\DSGE\\\\inputs_foreign.xlsx"))
}
\description{
This function imports world economic data from inputs_foreign file
}
