% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.rates_breakeven_inflation}
\alias{import.rates_breakeven_inflation}
\title{This function imports inflation expectation data from rates_breakeven_inflation}
\usage{

  import.rates_breakeven_inflation(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\",
  "Data\\\\BOI\\\\DSGE\\\\rates and ", "breakeven inflation.csv"))
}
\description{
This function imports inflation expectation data from rates_breakeven_inflation
}
