% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.real_rate_1Y}
\alias{import.real_rate_1Y}
\title{This function imports real short rate from real rate 1Y file}
\usage{

  import.real_rate_1Y(filepath = paste0("C:\\\\Users\\\\Misha\\\\Documents\\\\Data",
  "\\\\BOI\\\\DSGE\\\\real rate 1Y.csv"))
}
\description{
This function imports real short rate from real rate 1Y file
}
