% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{interpolate}
\alias{interpolate}
\title{This function interpolates forward by using linear interpolation}
\usage{
interpolate(data_vec, direction = "forward")
}
\arguments{
\item{data_vec}{numeric vector}

\item{direction}{direction for interpolation (default is "forward")}
}
\description{
This function interpolates forward by using linear interpolation
}
