% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_function.R
\name{pca_reduction}
\alias{pca_reduction}
\title{PCA reduction}
\usage{
pca_reduction(data, time_index = TRUE, scale = TRUE, n_comps = 1)
}
\arguments{
\item{data}{dataframe}

\item{time_index}{(optional) time index of the data}

\item{n_comps}{(optional) number of PCA components to return}
}
\description{
This function reduces dimension based on pca method
The function takes a data matrix and retuns the first n_comps
components of PCA transformation. If the data matrix has a
time index the results is aligned along the index
}
