% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimesion_reduction_functions.R
\name{pca_reduction}
\alias{pca_reduction}
\title{PCA reduction}
\usage{
pca_reduction(df, center = TRUE, scale = TRUE, sign_align_params = NULL)
}
\arguments{
\item{df}{dataframe}

\item{center}{boolean indicator}

\item{scale}{boolean indicator}

\item{sign_align_params}{(optional) a list of alignment parameters.
The first element in the list is the aligning (axis) variable, the value is either
character (variable's name) or numeric (variable's position index). The second element
if(supplied) is boolean indicator alignment direction (True means positive direction).}

\item{target_var_name}{string that specifies outcome feature}
}
\value{
a list with two elements : pca_obj - PCA object, time_index - Dates vector
}
\description{
This function reduces dimension based on PCA method
The function takes a data matrix and returns the first n_comps
components of PCA transformation. If the data matrix has a
time index the result is aligned along the index
}
